[CLEANUP] Renamed \fut\bot to \fut\api
Solves issue #14
