+++ src/JKetelaar/fut/api/errors/market/AmountTooBigException.php
@@ -3,7 +3,7 @@
  * @author JKetelaar
  */
 
-namespace JKetelaar\fut\bot\errors\market;
+namespace JKetelaar\fut\api\errors\market;
 
 class AmountTooBigException extends \Exception {
     /**
