+++ src/JKetelaar/fut/api/errors/market/UnknownEndpoint.php
@@ -3,7 +3,7 @@
  * @author JKetelaar
  */
 
-namespace JKetelaar\fut\bot\errors\market;
+namespace JKetelaar\fut\api\errors\market;
 
 class UnknownEndpoint extends \Exception {
     public function __construct($endpoint) {
