+++ src/JKetelaar/fut/api/market/items/ItemType.php
@@ -3,9 +3,9 @@
  * @author JKetelaar
  */
 
-namespace JKetelaar\fut\bot\market\items;
+namespace JKetelaar\fut\api\market\items;
 
-use JKetelaar\fut\bot\ImprovedEnum;
+use JKetelaar\fut\api\ImprovedEnum;
 
 class ItemType extends ImprovedEnum {
