+++ src/JKetelaar/fut/api/web/Parser.php
@@ -3,7 +3,7 @@
  * @author JKetelaar
  */
 
-namespace JKetelaar\fut\bot\web;
+namespace JKetelaar\fut\api\web;
 
 use simplehtmldom_1_5\simple_html_dom;
 use Sunra\PhpSimple\HtmlDomParser;
